refactor: batch updater

+++ internal/domain/repo/user.go
@@ -81,7 +81,7 @@ type UserRepository interface {
 	// UpdateUserUsedQuota 更新用户已用配额
 	UpdateUserUsedQuota(id int, quota int64) error
 	// UpdateUserRequestCount 更新用户请求计数
-	UpdateUserRequestCount(id int, count int) error
+	UpdateUserRequestCount(id int, count int64) error
 
 	// 软删除用户
 	DeleteUserById(id int) error
